+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.10.0-dev
+
+- Enhancements
+    - Databases are reloaded if a storage process gets restarted.
+    - Storage processes are now supervised.
+
 ## v0.9.0 (2015-11-16)
 
 - Enhancements